Share cats instances and test for Sum
