+++ core/src/test/scala/io/chrisdavenport/monoids/ProductTests.scala
@@ -7,8 +7,7 @@ import cats.laws.discipline._
 
 class ProductTests
   extends munit.DisciplineSuite
-    with MonoidsArbitraries
-    with ExtraMonoidsArbitraries {
+    with MonoidsArbitraries {
   checkAll("Product", EqTests[Product[Int]].eqv)
   checkAll("Product", OrderTests[Product[Int]].order)
   checkAll("Product", CommutativeMonoidTests[Product[Int]].commutativeMonoid)
