+++ core/src/test/scala/io/chrisdavenport/monoids/SumTests.scala
@@ -1,11 +1,11 @@
 package io.chrisdavenport.monoids
 
 import cats._
-import cats.implicits._
+import cats.syntax.all._
 import cats.kernel.laws.discipline._
 import cats.laws.discipline._
 
-class SumTests extends munit.DisciplineSuite with MonoidsArbitraries with ExtraMonoidsArbitraries {
+class SumTests extends munit.DisciplineSuite with MonoidsArbitraries {
   checkAll("Sum", EqTests[Sum[Int]].eqv)
   checkAll("Sum", OrderTests[Sum[Int]].order)
   checkAll("Sum", CommutativeMonoidTests[Sum[Int]].commutativeMonoid)
